fix(models): make NextAction.priority type match SQL (float, not int)
SQL column is REAL because storage._compute_next_action_priority uses (a.priority + b.priority) / 2 for midpoint reorders between two existing actions. The Pydantic int was being silently coerced on read, which worked but lied about the contract. Float reflects the real type and the field's actual role.

+++ backend/vellum/models.py
@@ -128,7 +128,9 @@ class NextAction(BaseModel):
     dossier_id: str
     action: str                           # short imperative
     rationale: str = ""
-    priority: int = 0                     # lower = higher priority; use spaced ints
+    priority: float = 0.0                 # lower = higher priority. REAL in SQL
+                                          # so midpoint reorders between two existing
+                                          # actions can store (a.priority + b.priority) / 2
     completed: bool = False
     completed_at: Optional[datetime] = None
     created_at: datetime
